+++ sdk/src/metrics/meter_provider.cc
@@ -1,6 +1,7 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
+#include <atomic>
 #include <chrono>
 #include <mutex>
 #include <utility>
@@ -143,6 +144,11 @@ void MeterProvider::SetExemplarFilter(metrics::ExemplarFilterType exemplar_filte
  */
 bool MeterProvider::Shutdown(std::chrono::microseconds timeout) noexcept
 {
+  // Shutdown only once
+  if (shutdown_latch_.test_and_set(std::memory_order_acquire))
+  {
+    return true;
+  }
   return context_->Shutdown(timeout);
 }
 
@@ -162,7 +168,7 @@ MeterProvider::~MeterProvider()
 {
   if (context_)
   {
-    context_->Shutdown();
+    Shutdown();
   }
 }